backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Web search and content extraction
duckduckgo-search>=4.1.0
trafilatura>=1.6.0
httpx>=0.26.0

# LangChain and OpenAI
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0

# Language detection
langdetect>=1.0.9

# Data validation
pydantic>=2.5.0

# Server-Sent Events
sse-starlette>=1.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
hypothesis>=6.92.0
